refactor: move sharing to dedicated module
