+++ unitycatalog/common/src/services/mod.rs
@@ -1,13 +1,11 @@
 use std::sync::Arc;
 
-use delta_kernel_datafusion::{TableSnapshot, Version};
+use deltalake_datafusion::{TableSnapshot, Version};
 
 use self::kernel::TableManager;
-use crate::api::{RequestContext, SharingQueryHandler};
-use crate::models::sharing::v1::*;
-use crate::models::tables::v1::{DataSourceFormat, TableInfo};
+use crate::models::tables::v1::DataSourceFormat;
 use crate::resources::ResourceStore;
-use crate::{ProvidesResourceStore, ResourceIdent, ResourceName, Result, ShareInfo};
+use crate::{ProvidesResourceStore, Result};
 
 pub mod kernel;
 mod location;
@@ -15,6 +13,7 @@ mod object_store;
 pub mod policy;
 pub mod secrets;
 pub mod session;
+mod sharing;
 
 pub use location::*;
 pub use policy::*;
@@ -110,78 +109,5 @@ impl TableManager for ServerHandler {
         version: Option<Version>,
     ) -> Result<Arc<dyn TableSnapshot>> {
         self.session.read_snapshot(location, format, version).await
-    }
-}
-
-#[async_trait::async_trait]
-trait SharingExt {
-    async fn get_snapshot(
-        &self,
-        share: &str,
-        schema: &str,
-        table: &str,
-    ) -> Result<Arc<dyn TableSnapshot>>;
-}
-
-#[async_trait::async_trait]
-impl<T: TableManager + ResourceStore> SharingExt for T {
-    async fn get_snapshot(
-        &self,
-        share: &str,
-        schema: &str,
-        table: &str,
-    ) -> Result<Arc<dyn TableSnapshot>> {
-        let share_ident = ResourceIdent::share(ResourceName::new([share]));
-        let share_info: ShareInfo = self.get(&share_ident).await?.0.try_into()?;
-        let Some(table_object) = share_info
-            .data_objects
-            .iter()
-            .find(|o| o.shared_as() == &format!("{}.{}", schema, table))
-        else {
-            return Err(crate::Error::NotFound);
-        };
-        let table_ident = ResourceIdent::table(ResourceName::new(table_object.name.split(".")));
-        let table_info: TableInfo = self.get(&table_ident).await?.0.try_into()?;
-        let location = table_info.storage_location.ok_or(crate::Error::NotFound)?;
-        let location = StorageLocationUrl::parse(&location)?;
-        self.read_snapshot(&location, &DataSourceFormat::Delta, None)
-            .await
-    }
-}
-
-#[async_trait::async_trait]
-impl SharingQueryHandler for ServerHandler {
-    async fn get_table_version(
-        &self,
-        request: GetTableVersionRequest,
-        context: RequestContext,
-    ) -> Result<GetTableVersionResponse> {
-        self.check_required(&request, context.recipient()).await?;
-        let snapshot = self
-            .get_snapshot(&request.share, &request.schema, &request.name)
-            .await?;
-        Ok(GetTableVersionResponse {
-            version: snapshot.version() as i64,
-        })
-    }
-
-    async fn get_table_metadata(
-        &self,
-        request: GetTableMetadataRequest,
-        context: RequestContext,
-    ) -> Result<QueryResponse> {
-        self.check_required(&request, context.recipient()).await?;
-        let snapshot = self
-            .get_snapshot(&request.share, &request.schema, &request.name)
-            .await?;
-        Ok([snapshot.metadata().into(), snapshot.protocol().into()].into())
-    }
-
-    async fn query_table(
-        &self,
-        request: QueryTableRequest,
-        context: RequestContext,
-    ) -> Result<QueryResponse> {
-        todo!()
     }
 }
